split parsers from core substitution codE
