+++ config_utilities/src/substitution_parsers.cpp
@@ -1,0 +1,72 @@
+/** -----------------------------------------------------------------------------
+ * Copyright (c) 2023 Massachusetts Institute of Technology.
+ * All Rights Reserved.
+ *
+ * AUTHORS:     Lukas Schmid <lschmid@mit.edu>, Nathan Hughes <na26933@mit.edu>
+ * AFFILIATION: MIT-SPARK Lab, Massachusetts Institute of Technology
+ * YEAR:        2023
+ * LICENSE:     BSD 3-Clause
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of its
+ * contributors may be used to endorse or promote products derived from
+ * this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ * -------------------------------------------------------------------------- */
+
+#include "config_utilities/substitution_parsers.h"
+
+#include <cstdlib>
+
+#include "config_utilities/internal/logger.h"
+
+namespace config {
+namespace {
+
+static const auto env_reg = RegisteredSubstitutions::Registration<EnvSubstitution>();
+static const auto var_reg = RegisteredSubstitutions::Registration<VarSubstitution>();
+
+}  // namespace
+
+std::string EnvSubstitution::process(const ParserContext& context, const std::string& contents) const {
+  const auto ret = std::getenv(contents.c_str());
+  if (!ret) {
+    context.error();
+    internal::Logger::logError("Failed to get envname from '" + contents + "'");
+    return contents;
+  }
+
+  return std::string(ret);
+}
+
+std::string VarSubstitution::process(const ParserContext& context, const std::string& contents) const {
+  auto iter = context.vars.find(contents);
+  if (iter == context.vars.end()) {
+    internal::Logger::logError("Unknown var '" + contents + "'");
+    context.error();
+    return contents;
+  }
+
+  return iter->second;
+}
+
+}  // namespace config
